backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncer==0.0.8
attrs==25.3.0
Authlib==1.6.0
backoff==2.2.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
databricks-sdk==0.57.0
datasets==3.6.0
deprecation==2.1.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
dspy==2.6.27
dspy-ai==2.6.27
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.14
fastapi-cli==0.0.7
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
grpcio==1.73.1
grpcio-health-checking==1.73.1
grpcio-tools==1.73.1
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json_repair==0.47.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
litellm==1.73.6
lxml==6.0.0
magicattr==0.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mlflow==3.1.1
mlflow-skinny==3.1.1
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
mailersend-python
narwhals==1.45.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
openai==1.93.0
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
optuna==4.4.0
packaging==25.0
pandas==2.3.0
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
PyPDF2==3.0.1
psycopg2-binary
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rich-toolkit==0.14.8
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==5.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.2
streamlit==1.46.1
sympy==1.14.0
tavily-python==0.7.9
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
validators==0.34.0
watchfiles==1.1.0
weaviate-client==4.15.4
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
